+++ main.py
@@ -1,14 +1,10 @@
-users:list=(
-    {"name":"Zosia", "location":"Poznań","posts":400},
-    {"name":"Monika", "location":"Nowy_Sącz","posts":500},
-    {"name":"Julka", "location":"Rzeszów","posts":700},
-    {"name":"Beata", "location":"Lublin","posts":200},
-)
+from utils.model import users
+from utils.controller import get_user_info
 
 
+def main():
+    get_user_info(users)
 
-def get_user_info(users_data:list)->None:
-    for user in users_data:
-        print(f"Twój znajomy {user['name']} z miejscowości {user['location']} opublikował {user['posts']} postów")
 
-get_user_info(users)
+if __name__ == "__main__":
+    main()
