+++ utils/model.py
@@ -1,0 +1,6 @@
+users:list=(
+    {"name":"Zosia", "location":"Poznań","posts":400},
+    {"name":"Monika", "location":"Nowy_Sącz","posts":500},
+    {"name":"Julka", "location":"Rzeszów","posts":700},
+    {"name":"Beata", "location":"Lublin","posts":200},
+)
